requirements.txt
Flask==3.0.0
twilio==9.0.0
python-dotenv==1.0.0
spacy==3.7.2
transformers==4.35.0
torch==2.1.0
chromadb==0.4.15
sentence-transformers==2.2.2
redis==5.0.1